+++ test/simple_sum.sol
@@ -1,7 +1,7 @@
 package example
 func main() -> int {
     values: List<int> = [10, 20]
-    mut total: long = 0
+    mut total: int = 0
     for value in values {
         total = total + value
     }
